requirements.txt
streamlit
psycopg2-binary
openai
python-dotenv
pytest
pre-commit
pillow
